+++ src/main/java/com/clubber/ClubberServer/global/infrastructure/outer/discord/client/DiscordClient.java
@@ -1,7 +1,7 @@
-package com.clubber.ClubberServer.global.infrastructure.outer.exception.client.discord;
+package com.clubber.ClubberServer.global.infrastructure.outer.discord.client;
 
 import com.clubber.ClubberServer.global.config.feign.FeignConfig;
-import com.clubber.ClubberServer.global.infrastructure.outer.exception.dto.discord.DiscordMessage;
+import com.clubber.ClubberServer.global.infrastructure.outer.discord.dto.discord.DiscordMessage;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.PostMapping;
